requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy
tqdm
tensorboard
matplotlib
pandas

datasets
editdistance
wandb
pillow

sentencepiece
wandb

requests==2.31.0
platformdirs==3.11.0
psutil==5.8
decorator==4.4.2

photosynthesis-team::piq  
